Add files via upload
This project was developed as part of PwC's PowerBI Job Simulation through the Forage platform. The aim of the simulation was to emulate real-world business scenarios and tasks that a data analyst might encounter in a professional setting, focusing on leveraging Power BI for data visualization and business intelligence.
*Tools and Technologies
Power BI: For data visualization and dashboard creation.
Power Query: For data cleaning and transformation.
DAX (Data Analysis Expressions): For creating custom calculations and measures.